@@ -3,15 +3,9 @@
 namespace App\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
+use MongoDB\Laravel\Eloquent\Model;
 
 class Role extends Model
 {
     use HasFactory;
-
-    protected $guarded = [];
-
-    public function users() {
-        return $this->hasMany(User::class);
-    }
 }
