@@ -2,38 +2,25 @@
 
 namespace App\Providers;
 
+// use Illuminate\Support\Facades\Gate;
 use Illuminate\Foundation\Support\Providers\AuthServiceProvider as ServiceProvider;
-use Illuminate\Support\Facades\Gate;
 
 class AuthServiceProvider extends ServiceProvider
 {
     /**
-     * The policy mappings for the application.
+     * The model to policy mappings for the application.
      *
      * @var array<class-string, class-string>
      */
     protected $policies = [
-        Tags::class => TagPolicy::class,
-        User::class => UserPolicy::class,
-        Item::class => ItemPolicy::class,
-        Role::class => RolePolicy::class,
-        Category::class => CategoryPolicy::class,
+        //
     ];
 
     /**
      * Register any authentication / authorization services.
-     *
-     * @return void
      */
-    public function boot()
+    public function boot(): void
     {
-        $this->registerPolicies();
-
-        Gate::define('manage-items', 'App\Policies\UserPolicy@manageItems');
-
-        Gate::define('manage-users', 'App\Policies\UserPolicy@manageUsers');
-
-        Gate::define('manage-users', 'App\Policies\UserPolicy@updatecategory');
-
+        //
     }
 }
