@@ -56,20 +56,6 @@ return [
             'throw' => false,
         ],
 
-        'avatars' => [
-            'driver' => 'local',
-            'root' => storage_path('app/public/avatars'),
-            'url' => env('APP_URL').'/avatars',
-            'visibility' => 'public',
-        ],
-
-        'items' => [
-            'driver' => 'local',
-            'root' => storage_path('app/public/items'),
-            'url' => env('APP_URL').'/items',
-            'visibility' => 'public',
-        ],
-
     ],
 
     /*
@@ -84,9 +70,7 @@ return [
     */
 
     'links' => [
-        // public_path('storage') => storage_path('app/public'),
-        public_path('avatars') => storage_path('app/public/avatars'),
-        public_path('items') => storage_path('app/public/items'),
+        public_path('storage') => storage_path('app/public'),
     ],
 
 ];
