@@ -1,5 +1,1 @@
-import PerfectScrollbar from 'perfect-scrollbar';
-window.PerfectScrollbar = PerfectScrollbar;
-
-require('./bootstrap');
-require('./custom')
+import './bootstrap';
